% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/viaf_suggest.R
\name{viaf_suggest}
\alias{viaf_suggest}
\title{Suggest Authority Terms based on a text passed in a query}
\usage{
viaf_suggest(query = NULL)
}
\arguments{
\item{query}{search term}
}
\description{
This function searches for VIAF entries
}
\examples{
dontrun{
viaf_suggest("Egelhaaf, Martin")}
}

